requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.5.2
python-dotenv==1.0.1
qrcode==8.2
<<<<<<< HEAD
pillow==12.0.0
=======
pillow
>>>>>>> 73227df581c5d6fff1eaaaa50cb7a0ee5613b151
